finish containerize for read_apis, update CI workflow
