+++ read_apis/Dockerfile
@@ -9,20 +9,10 @@ COPY go.sum ./
 
 RUN go mod download
 
-# copy individually so that if there is cache of any file/folder, line can be skipped
-# COPY run ./
-# COPY url.txt ./
-# COPY cmd ./cmd
-# COPY github_apis ./github_apis
-# COPY internal ./internal
-# COPY tests ./tests
-
 COPY . /app/
-
-# unsure how to build this app
-# RUN go build -o /package_rater_app
 
 # for web app purposes
 EXPOSE 8080
 
 # CMD ["/package_rater_app"]
+ENTRYPOINT [ "go", "run", "src/main.go" ]
