+++ read_apis/src/read_test.go
@@ -1,4 +1,4 @@
-package src
+package main
 
 import (
 	"testing"
